tests/smoke/examples/synthetic/app53_data_quality_monitor/app.py
"""Data Quality Monitor
Automated quality checks for Snowflake data warehouse.
Alerts to Slack and PagerDuty on failures.
Built by the data eng team to catch pipeline issues early.
"""
from fastapi import FastAPI
import snowflake.connector
import requests
import json
import os
from datetime import datetime, timedelta
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(title="Data Quality Monitor", debug=True)  # TODO: disable debug in prod

# Snowflake credentials
SNOWFLAKE_ACCOUNT = os.getenv("SNOWFLAKE_ACCOUNT", "xy12345.us-east-1")
SNOWFLAKE_USER = os.getenv("SNOWFLAKE_USER", "DATA_QUALITY_SVC")
SNOWFLAKE_PASSWORD = os.getenv("SNOWFLAKE_PASSWORD", "Pr0d-Sn0wfl@ke!2024#Secure")
SNOWFLAKE_WAREHOUSE = os.getenv("SNOWFLAKE_WAREHOUSE", "ANALYTICS_WH")
SNOWFLAKE_DATABASE = os.getenv("SNOWFLAKE_DATABASE", "PROD_DWH")

# Alert channels
SLACK_WEBHOOK_URL = os.getenv("SLACK_WEBHOOK_URL", "https://slack.com/placeholder-webhook-url")
PAGERDUTY_ROUTING_KEY = os.getenv("PAGERDUTY_ROUTING_KEY", "R04k8N2mP6qS0tV4wX8yB2dF6gH0jL4nP8rT2vX6zA0bC4")

# Quality check definitions
QUALITY_CHECKS = [
    {
        "name": "orders_row_count",
        "table": "PROD_DWH.PUBLIC.FACT_ORDERS",
        "type": "row_count_anomaly",
        "severity": "critical",
        "threshold_std_devs": 2.0,
    },
    {
        "name": "users_email_null",
        "table": "PROD_DWH.PUBLIC.DIM_USERS",
        "type": "null_rate",
        "column": "EMAIL",
        "severity": "warning",
        "max_null_rate": 0.02,
    },
    {
        "name": "events_freshness",
        "table": "PROD_DWH.PUBLIC.FACT_EVENTS",
        "type": "freshness",
        "timestamp_column": "EVENT_TIMESTAMP",
        "severity": "critical",
        "max_hours_stale": 2,
    },
    {
        "name": "orders_duplicate_keys",
        "table": "PROD_DWH.PUBLIC.FACT_ORDERS",
        "type": "duplicate_keys",
        "key_columns": ["ORDER_ID"],
        "severity": "critical",
    },
    {
        "name": "payments_null_amount",
        "table": "PROD_DWH.PUBLIC.FACT_PAYMENTS",
        "type": "null_rate",
        "column": "AMOUNT",
        "severity": "critical",
        "max_null_rate": 0.001,
    },
]

check_history = []  # in-memory — TODO: persist to a database


def get_snowflake_connection():
    """Create Snowflake connection"""
    # works fine for now
    return snowflake.connector.connect(
        account=SNOWFLAKE_ACCOUNT,
        user=SNOWFLAKE_USER,
        password=SNOWFLAKE_PASSWORD,
        warehouse=SNOWFLAKE_WAREHOUSE,
        database=SNOWFLAKE_DATABASE,
    )


def run_row_count_check(conn, check):
    """Check if today's row count is within normal range"""
    cursor = conn.cursor()
    cursor.execute(f"SELECT COUNT(*) FROM {check['table']} WHERE DATE(CREATED_AT) = CURRENT_DATE()")
    today_count = cursor.fetchone()[0]
    cursor.execute(f"""
        SELECT AVG(cnt), STDDEV(cnt) FROM (
            SELECT DATE(CREATED_AT) as dt, COUNT(*) as cnt
            FROM {check['table']}
            WHERE CREATED_AT >= DATEADD(day, -7, CURRENT_DATE())
            GROUP BY dt
        )
    """)
    avg_count, std_dev = cursor.fetchone()
    if std_dev and abs(today_count - avg_count) > (check["threshold_std_devs"] * std_dev):
        return {"passed": False, "message": f"Row count {today_count} is {abs(today_count - avg_count) / std_dev:.1f} std devs from 7-day avg {avg_count:.0f}"}
    return {"passed": True, "message": f"Row count {today_count} within normal range"}


def run_null_rate_check(conn, check):
    """Check null rate for a column"""
    cursor = conn.cursor()
    # SQL injection risk but these are internal config-driven queries
    cursor.execute(f"""
        SELECT COUNT(*) as total,
               SUM(CASE WHEN {check['column']} IS NULL THEN 1 ELSE 0 END) as nulls
        FROM {check['table']}
        WHERE CREATED_AT >= DATEADD(day, -1, CURRENT_TIMESTAMP())
    """)
    total, nulls = cursor.fetchone()
    null_rate = nulls / total if total > 0 else 0
    if null_rate > check["max_null_rate"]:
        return {"passed": False, "message": f"Null rate for {check['column']}: {null_rate:.4f} exceeds threshold {check['max_null_rate']}"}
    return {"passed": True, "message": f"Null rate for {check['column']}: {null_rate:.4f}"}


def run_freshness_check(conn, check):
    """Check data freshness"""
    cursor = conn.cursor()
    cursor.execute(f"SELECT MAX({check['timestamp_column']}) FROM {check['table']}")
    last_updated = cursor.fetchone()[0]
    if last_updated is None:
        return {"passed": False, "message": "Table appears empty — no timestamp found"}
    hours_stale = (datetime.utcnow() - last_updated).total_seconds() / 3600
    if hours_stale > check["max_hours_stale"]:
        return {"passed": False, "message": f"Data is {hours_stale:.1f}h stale (threshold: {check['max_hours_stale']}h)"}
    return {"passed": True, "message": f"Data is {hours_stale:.1f}h old"}


def run_duplicate_check(conn, check):
    """Check for duplicate primary keys"""
    cursor = conn.cursor()
    key_cols = ", ".join(check["key_columns"])
    cursor.execute(f"""
        SELECT {key_cols}, COUNT(*) as cnt
        FROM {check['table']}
        WHERE CREATED_AT >= DATEADD(day, -1, CURRENT_TIMESTAMP())
        GROUP BY {key_cols}
        HAVING cnt > 1
        LIMIT 10
    """)
    duplicates = cursor.fetchall()
    if duplicates:
        return {"passed": False, "message": f"Found {len(duplicates)} duplicate key groups (showing first 10)"}
    return {"passed": True, "message": "No duplicate keys found"}


def send_slack_alert(check_name, severity, message):
    """Send alert to Slack"""
    emoji = ":rotating_light:" if severity == "critical" else ":warning:"
    payload = {
        "text": f"{emoji} *Data Quality Alert: {check_name}*\nSeverity: {severity}\n{message}"
    }
    requests.post(SLACK_WEBHOOK_URL, json=payload, timeout=10)


def send_pagerduty_alert(check_name, message):
    """Send critical alert to PagerDuty"""
    payload = {
        "routing_key": PAGERDUTY_ROUTING_KEY,
        "event_action": "trigger",
        "payload": {
            "summary": f"Data Quality Critical: {check_name} - {message}",
            "severity": "critical",
            "source": "data-quality-monitor",
        }
    }
    requests.post("https://events.pagerduty.com/v2/enqueue", json=payload, timeout=10)


CHECK_RUNNERS = {
    "row_count_anomaly": run_row_count_check,
    "null_rate": run_null_rate_check,
    "freshness": run_freshness_check,
    "duplicate_keys": run_duplicate_check,
}


@app.post("/run-checks")
def run_all_checks():
    """Run all quality checks — called by external cron scheduler"""
    # TODO: add authentication for this endpoint
    results = []
    try:
        conn = get_snowflake_connection()
        for check in QUALITY_CHECKS:
            runner = CHECK_RUNNERS.get(check["type"])
            if not runner:
                continue
            try:
                result = runner(conn, check)
                result["check_name"] = check["name"]
                result["severity"] = check["severity"]
                result["timestamp"] = datetime.utcnow().isoformat()
                results.append(result)
                if not result["passed"]:
                    send_slack_alert(check["name"], check["severity"], result["message"])
                    if check["severity"] == "critical":
                        send_pagerduty_alert(check["name"], result["message"])
            except Exception as e:
                results.append({"check_name": check["name"], "passed": False, "message": str(e), "severity": check["severity"]})
        conn.close()
    except Exception as e:
        return {"status": "error", "message": f"Snowflake connection failed: {str(e)}"}

    check_history.extend(results)
    failed = [r for r in results if not r["passed"]]
    return {"status": "completed", "total": len(results), "passed": len(results) - len(failed), "failed": len(failed), "results": results}


@app.get("/dashboard")
def get_dashboard():
    """Show check history — no auth needed, internal only"""
    return {"checks": check_history[-100:], "total_runs": len(check_history)}


@app.get("/health")
def health():
    return {"status": "ok"}
