tests/smoke/examples/synthetic/app53_data_quality_monitor/requirements.txt
fastapi==0.109.0
uvicorn
snowflake-connector-python==3.6.0
requests
python-dotenv==1.0.0
